Move Ready to the common crate
