+++ mio-common/src/lib.rs
@@ -5,5 +5,7 @@
 // for the `Token`.
 
 mod token;
+mod ready;
 
 pub use token::Token;
+pub use ready::Ready;
