+++ src/lib.rs
@@ -121,10 +121,10 @@ pub mod net;
 #[cfg_attr(not(feature = "doc-sys"), doc(hidden))]
 pub mod sys;
 
-pub use event_imp::{Interests, PollOpt, Ready};
+pub use event_imp::{Interests, PollOpt};
 pub use poll::{Poll, Registration, Registry, SetReadiness};
 
-pub use mio_common::Token;
+pub use mio_common::{Ready, Token};
 
 pub mod event {
     //! Readiness event types and utilities.
